+++ cogs/moderator.py
@@ -4,11 +4,11 @@ from discord import TextChannel, User, Member, Embed
 from discord.ext.commands import Cog, Bot, command, Context, cooldown, BucketType
 from discord.ext.tasks import loop
 
-from cogs.botStatus import listener
+from cogs.bot_status import listener
 from cogs.util.ainit_ctx_mgr import AinitManager
 from cogs.util.assign_variables import assign_role, assign_accepted_chats, assign_chat
 from cogs.util.placeholder import Placeholder
-from core.globalEnum import ConfigurationNameEnum
+from core.global_enum import ConfigurationNameEnum
 from core.logger import get_discord_child_logger
 from core.predicates import bot_chat, has_role_plus
