+++ cogs/util/study_subject_util.py
@@ -1,9 +1,9 @@
 from discord import CategoryChannel, Role, PermissionOverwrite, TextChannel, Guild, Embed
 from discord.ext.commands import Context
 
-from core.globalEnum import CollectionEnum, ConfigurationNameEnum
-from mongo.primitiveMongoData import PrimitiveMongoData
-from mongo.subjectsorgroups import SubjectOrGroup, SubjectsOrGroups
+from core.global_enum import CollectionEnum, ConfigurationNameEnum
+from mongo.primitive_mongo_data import PrimitiveMongoData
+from mongo.subjects_or_groups import SubjectOrGroup, SubjectsOrGroups
 
 
 class StudySubjectUtil:
