+++ core/global_enum.py
